Add another "Add item" button near the top
When a list is already quite long, it is not obvious how to add a new
item to the list. Add another "Add item" button near the top as a hint
for users on how to add new items.

Scroll to the full form when adding new items

When the "Add item" button is clicked, bring the full form into view.
No longer focus the input button, but the whole form to avoid the
keyboard popup on mobile devices.